+++ dali/pipeline/operator/common.h
@@ -1,4 +1,4 @@
-// Copyright (c) 2017-2021, NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+// Copyright (c) 2017-2022, NVIDIA CORPORATION & AFFILIATES. All rights reserved.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -24,6 +24,7 @@
 #include "dali/core/tensor_shape_print.h"
 #include "dali/pipeline/data/views.h"
 #include "dali/pipeline/operator/op_spec.h"
+#include "dali/pipeline/operator/sequence_info.h"
 #include "dali/pipeline/workspace/workspace.h"
 
 namespace dali {
@@ -31,17 +32,18 @@ template <typename T>
 inline void GetSingleOrRepeatedArg(const OpSpec &spec, vector<T> &result,
                                    const std::string &argName, size_t repeat_count = 2) {
   if (!spec.TryGetRepeatedArgument<T>(result, argName)) {
-      T scalar = spec.GetArgument<T>(argName);
-      result.assign(repeat_count, scalar);
+    T scalar = spec.GetArgument<T>(argName);
+    result.assign(repeat_count, scalar);
   } else if (result.size() == 1 && repeat_count != 1) {
-      T scalar = result.front();
-      result.assign(repeat_count, scalar);
+    T scalar = result.front();
+    result.assign(repeat_count, scalar);
   }
 
-  DALI_ENFORCE(result.size() == repeat_count,
-      "Argument \"" + argName + "\" expects either a single value "
-      "or a list of " + to_string(repeat_count) + " elements. " +
-      to_string(result.size()) + " given.");
+  DALI_ENFORCE(result.size() == repeat_count, "Argument \"" + argName +
+                                                  "\" expects either a single value "
+                                                  "or a list of " +
+                                                  to_string(repeat_count) + " elements. " +
+                                                  to_string(result.size()) + " given.");
 }
 
 template <typename T>
@@ -66,12 +68,10 @@ void GetPerSampleArgument(std::vector<T> &output, const std::string &argument_na
         output[i] = data[i];
       }
     } else {
-      bool is_valid_shape = N == batch_size &&
-                            is_uniform(shape) &&
-                            volume(shape.tensor_shape_span(0)) == 1;
-      DALI_ENFORCE(is_valid_shape,
-        make_string("`", argument_name, "` must be a 1xN or Nx1 (N = ", batch_size,
-                    ") tensor list. Got: ", shape));
+      bool is_valid_shape =
+          N == batch_size && is_uniform(shape) && volume(shape.tensor_shape_span(0)) == 1;
+      DALI_ENFORCE(is_valid_shape, make_string("`", argument_name, "` must be a 1xN or Nx1 (N = ",
+                                               batch_size, ") tensor list. Got: ", shape));
 
       output.resize(batch_size);
       for (int i = 0; i < batch_size; i++) {
@@ -98,23 +98,23 @@ void GetPerSampleArgument(std::vector<T> &output, const std::string &argument_na
  * TODO(michalz): rework it somehow to avoid going through this logic for each sample
  */
 template <typename T>
-void GetGeneralizedArg(span<T> result, const std::string &name, int sample_idx, const OpSpec& spec,
-                       const ArgumentWorkspace& ws) {
+void GetGeneralizedArg(span<T> result, const std::string &name, int sample_idx, const OpSpec &spec,
+                       const ArgumentWorkspace &ws, const SampleFrameInfoFn &fr_info = {}) {
   int argument_length = result.size();
   if (spec.HasTensorArgument(name)) {
-    const auto& tv = ws.ArgumentInput(name);
-    const auto& tensor = tv[sample_idx];
-    const auto& shape = tensor.shape();
+    const auto &tv = ws.ArgumentInput(name);
+    const auto &tensor = tv[sample_idx];
+    const auto &shape = tensor.shape();
     auto vol = volume(shape);
     if (shape.size() != 0) {
-      DALI_ENFORCE(shape.size() == 1,
-                   make_string("Argument ", name, " for sample ", sample_idx,
-                               " is expected to be a scalar or a 1D tensor, got: ",
-                               shape.size(), "D."));
-      DALI_ENFORCE(vol == 1 || vol == argument_length,
-                   make_string("Argument ", name, " for sample ", sample_idx,
-                               " is expected to have 1 or ", argument_length,
-                               " elements, got: ", shape[0], "."));
+      DALI_ENFORCE(
+          shape.size() == 1,
+          make_string("Argument ", name, " for ", fr_info(sample_idx),
+                      " is expected to be a scalar or a 1D tensor, got: ", shape.size(), "D."));
+      DALI_ENFORCE(
+          vol == 1 || vol == argument_length,
+          make_string("Argument ", name, " for ", fr_info(sample_idx), " is expected to have 1 or ",
+                      argument_length, " elements, got: ", shape[0], "."));
     }
     if (vol == 1) {
       for (int i = 0; i < argument_length; i++) {
@@ -140,9 +140,9 @@ std::pair<int, int> GetShapeLikeArgument(std::vector<ExtentType> &out_shape, con
                make_string("Invalid batch size. Expected nonnegative, actual: ", batch_size));
 
   auto to_extent = [](ArgumentType extent) {
-    return std::is_floating_point<ArgumentType>::value && std::is_integral<ExtentType>::value
-               ? static_cast<ExtentType>(std::lround(extent))
-               : static_cast<ExtentType>(extent);
+    return std::is_floating_point<ArgumentType>::value && std::is_integral<ExtentType>::value ?
+               static_cast<ExtentType>(std::lround(extent)) :
+               static_cast<ExtentType>(extent);
   };
 
   if (spec.HasTensorArgument(argument_name)) {
@@ -150,11 +150,12 @@ std::pair<int, int> GetShapeLikeArgument(std::vector<ExtentType> &out_shape, con
     auto argview = view<const ArgumentType>(arg);
     int N = argview.shape.num_samples();
     DALI_ENFORCE(N == batch_size, make_string("Unexpected number of samples in argument `",
-      argument_name, "` (expected: ", batch_size, ")"));
-    DALI_ENFORCE(is_uniform(argview.shape), "A tensor list shape must have the same dimensionality "
-      "for all samples.");
+                                              argument_name, "` (expected: ", batch_size, ")"));
+    DALI_ENFORCE(is_uniform(argview.shape),
+                 "A tensor list shape must have the same dimensionality "
+                 "for all samples.");
     if (argview.shape.sample_dim() == 0) {  // a list of true scalars (0D)
-      if (ndim < 0)  // no ndim? assume 1D
+      if (ndim < 0)                         // no ndim? assume 1D
         ndim = 1;
       out_shape.resize(N * ndim);
       for (int i = 0; i < N; i++) {
@@ -163,11 +164,12 @@ std::pair<int, int> GetShapeLikeArgument(std::vector<ExtentType> &out_shape, con
           out_shape[i * ndim + d] = e;  // broadcast scalar size to all dims
       }
     } else {
-      DALI_ENFORCE(argview.shape.sample_dim() == 1, "Shapes must be 1D tensors with extent equal "
-       "to shape dimensionality (or scalar)");
+      DALI_ENFORCE(argview.shape.sample_dim() == 1,
+                   "Shapes must be 1D tensors with extent equal "
+                   "to shape dimensionality (or scalar)");
       int D = argview.shape[0][0];
-      DALI_ENFORCE(ndim < 0 || D == ndim, make_string(D, "-element tensor cannot describe an ",
-        ndim, "D shape."));
+      DALI_ENFORCE(ndim < 0 || D == ndim,
+                   make_string(D, "-element tensor cannot describe an ", ndim, "D shape."));
       ndim = D;
       out_shape.resize(N * D);
       for (int i = 0; i < N; i++)
