+++ dali/pipeline/operator/sequence_workspace.h
@@ -1,0 +1,54 @@
+// Copyright (c) 2022, NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#ifndef DALI_PIPELINE_OPERATOR_SEQUENCE_WORKSPACE_H_
+#define DALI_PIPELINE_OPERATOR_SEQUENCE_WORKSPACE_H_
+
+#include "dali/pipeline/operator/common.h"
+#include "dali/pipeline/operator/sequence_info.h"
+#include "dali/pipeline/util/backend2workspace_map.h"
+
+
+namespace dali {
+
+template <typename Backend>
+class SequenceWorkspaceView : public workspace_t<Backend> {
+ public:
+  void SetFrameInfoFns(std::vector<SampleFrameInfoFn> &&input_sample_info) {
+    input_sample_info_ = input_sample_info;
+  }
+
+  const SampleFrameInfoFn GetFrameInfoForInput(int input_idx) const {
+    return input_sample_info_[input_idx];
+  }
+
+ private:
+  std::vector<SampleFrameInfoFn> input_sample_info_;
+};
+
+
+// TODO(ktokarski) Consider making it a method of workspace/argument workspace instead of
+// dynamic_cast. For now it seems too specific for a sequence operator context to place it there
+template <typename Backend>
+const SampleFrameInfoFn GetFrameInfoForInput(const workspace_t<Backend> &ws, int input_idx) {
+  auto ws_ptr = dynamic_cast<const SequenceWorkspaceView<Backend> *>(&ws);
+  if (!ws_ptr) {
+    return {};
+  }
+  return ws_ptr->GetFrameInfoForInput(input_idx);
+}
+
+}  // namespace dali
+
+#endif  // DALI_PIPELINE_OPERATOR_SEQUENCE_WORKSPACE_H_
